HuggingFace/requirements.txt
streamlit
tensorflow==2.18.0
numpy
Pillow
jupyter

